requirements.txt
python-dotenv
pymongo
Flask
deepface
tf-keras
opencv-python
numpy
Pillow
